+++ core/src/components/range/test/a11y/range.e2e.ts
@@ -33,7 +33,7 @@ configs({ modes: ['md'], directions: ['ltr'] }).forEach(({ title, screenshot, co
       await page.setContent(
         `<ion-app>
           <ion-content>
-            <ion-range min="0" max="100" value="80" legacy="true"></ion-range>
+            <ion-range min="0" max="100" value="80" aria-label="Range"></ion-range>
           </ion-content>
         </ion-app>
           `,
@@ -67,7 +67,7 @@ configs({ directions: ['ltr'] }).forEach(({ title, screenshot, config }) => {
       await page.setContent(
         `<ion-app>
           <ion-content>
-            <ion-range min="0" max="100" value="50" pin="true" legacy="true"></ion-range>
+            <ion-range min="0" max="100" value="50" pin="true" aria-label="Range"></ion-range>
           </ion-content>
         </ion-app>
         `,